Standarized Tags to REST
